backend/requirements.txt
Flask==2.2.5
flask-cors
psycopg2-binary
